style(rust): green the workspace clippy -D warnings baseline
Clear all ~124 workspace clippy -D warnings errors so CLI CI's
`cargo clippy --locked --tests -- -D warnings` passes. Mechanical, non-behavioral:

- cargo clippy --fix bulk: unused imports, collapsible-if/let-chains, needless
  borrows, manual RangeInclusive::contains, useless vec!/conversions,
  iter().any() over contains(), map_or, ?/Some sugar, auto-deref.
- Targeted #[allow] for intentional design lints: too_many_arguments,
  should_implement_trait (inherent from_str), large_enum_variant (frozen wire
  enums), result_unit_err, await_holding_lock (single-threaded test executor).
- Small non-behavioral edits: is_empty helper on MemoryKekStore, a type alias
  for a complex coordinator type, &mut Vec -> &mut [_] on a sort-only fn,
  struct-init over mutate-default, doc-list indentation, collapse of two
  genuinely-identical if/else arms, removal of a dead re-assignment.
- Restored a handful of items the --fix stripped that are only used by
  #[cfg(test)] code, re-gated under #[cfg(test)].
- #[allow(dead_code)] on never-used items to preserve intended scaffolding.

NOTE — dead-code allows hide three clusters that are test-covered but have NO
production caller and may be intended-but-unwired paths, worth a follow-up:
  * secret_ref.rs: migrate_effective_layers_once (zero callers),
    migrate_provider_files, load_paths_with_secret_migration
  * redact/protected_redaction_history.rs: RedactionRehydrationFrame + its
    methods, append_and_attach, rehydrate_for_artifact, fingerprint

fmt is clean; clippy --workspace --tests -D warnings exits 0; no NEW test
failures vs the (large, pre-existing) nextest baseline.

+++ crates/cockpit-core/src/private_fs.rs
@@ -643,7 +643,7 @@ fn walk_private_dir(path: &Path, create: bool) -> Result<std::fs::File, PrivateF
                 }
                 // SAFETY: `fstatat` returned 0, so `stat` is initialised.
                 let stat = unsafe { stat.assume_init() };
-                if u32::from(stat.st_mode) & u32::from(libc::S_IFMT) != u32::from(libc::S_IFLNK) {
+                if stat.st_mode & libc::S_IFMT != libc::S_IFLNK {
                     return Err(PrivateFsError::Containment(format!(
                         "component {name:?} under {} is not a directory",
                         path.display()
@@ -1151,20 +1151,20 @@ fn refuse_hostile_target(
     }
     // SAFETY: `fstatat` returned 0, so `stat` is initialised.
     let stat = unsafe { stat.assume_init() };
-    let kind = u32::from(stat.st_mode) & u32::from(libc::S_IFMT);
-    if kind == u32::from(libc::S_IFLNK) {
+    let kind = stat.st_mode & libc::S_IFMT;
+    if kind == libc::S_IFLNK {
         return Err(PrivateFsError::Containment(format!(
             "{}: write target is a symlink",
             label.display()
         )));
     }
-    if kind != u32::from(libc::S_IFREG) {
+    if kind != libc::S_IFREG {
         return Err(PrivateFsError::Containment(format!(
             "{}: write target is not a regular file",
             label.display()
         )));
     }
-    if u64::from(stat.st_nlink) != 1 {
+    if stat.st_nlink != 1 {
         return Err(PrivateFsError::MultiplyLinked(format!(
             "{}: write target has {} hard links",
             label.display(),
@@ -1287,7 +1287,7 @@ fn write_private_file_unix(path: &Path, bytes: &[u8], publish: PrivateWritePubli
         let matches = stat_ok && {
             let named = unsafe { named.assume_init() };
             // `as u64`: `dev_t`/`ino_t` widths differ across Unix targets.
-            named.st_dev as u64 == held.dev() && named.st_ino as u64 == held.ino()
+            named.st_dev == held.dev() && named.st_ino == held.ino()
         };
         if !matches {
             unlinkat_best_effort(&dir_handle, &temp_c);
